More refactoring; rename optional fields component
